update drone config

+++ test/e2e/e2e_suite_test.go
@@ -3,6 +3,8 @@ package e2e
 import (
 	"context"
 	"fmt"
+	"os"
+	"strconv"
 	"testing"
 
 	. "github.com/onsi/ginkgo"
@@ -19,11 +21,11 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/utils/pointer"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 )
 
 func TestE2E(t *testing.T) {
@@ -39,6 +41,7 @@ var (
 	testEnv     *envtest.Environment
 	k8sClient   crclient.Client
 	restConfig  *rest.Config
+	useExisting bool
 )
 
 func deleteTestClusterIfExists() {
@@ -55,47 +58,57 @@ func deleteTestClusterIfExists() {
 var _ = BeforeSuite(func() {
 	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
 
-	deleteTestClusterIfExists()
-	freePort, err := freeport.GetFreePort()
-	Expect(err).NotTo(HaveOccurred())
-	simpleConfig := v1alpha2.SimpleConfig{
-		Name:    clusterName,
-		Image:   "rancher/k3s:latest",
-		Servers: 1,
-		Agents:  1,
-		ExposeAPI: v1alpha2.SimpleExposureOpts{
-			HostIP:   "127.0.0.1",
-			HostPort: fmt.Sprint(freePort),
-		},
-		Ports: []v1alpha2.PortWithNodeFilters{
-			{
-				Port:        "8081:80",
-				NodeFilters: []string{"loadbalancer"},
-			},
-		},
-		Options: v1alpha2.SimpleConfigOptions{
-			K3sOptions: v1alpha2.SimpleConfigOptionsK3s{
-				ExtraServerArgs: []string{
-					"--log=/var/log/k3s.log",
-					"--alsologtostderr",
-				},
-			},
-		},
+	if str, ok := os.LookupEnv("E2E_USE_EXISTING"); ok {
+		if value, err := strconv.ParseBool(str); err == nil {
+			useExisting = value
+		}
 	}
 
-	ctx := context.Background()
-	conf, err := config.TransformSimpleToClusterConfig(
-		ctx, runtimes.Docker, simpleConfig)
-	Expect(err).NotTo(HaveOccurred())
+	if !useExisting {
+		deleteTestClusterIfExists()
+		freePort, err := freeport.GetFreePort()
+		Expect(err).NotTo(HaveOccurred())
+		simpleConfig := v1alpha2.SimpleConfig{
+			Name:    clusterName,
+			Image:   "rancher/k3s:latest",
+			Servers: 1,
+			Agents:  1,
+			ExposeAPI: v1alpha2.SimpleExposureOpts{
+				HostIP:   "127.0.0.1",
+				HostPort: fmt.Sprint(freePort),
+			},
+			Ports: []v1alpha2.PortWithNodeFilters{
+				{
+					Port:        "8081:80",
+					NodeFilters: []string{"loadbalancer"},
+				},
+			},
+			Options: v1alpha2.SimpleConfigOptions{
+				K3sOptions: v1alpha2.SimpleConfigOptionsK3s{
+					ExtraServerArgs: []string{
+						"--log=/var/log/k3s.log",
+						"--alsologtostderr",
+					},
+				},
+			},
+		}
 
-	err = client.ClusterRun(ctx, runtimes.Docker, conf)
-	Expect(err).NotTo(HaveOccurred())
+		ctx := context.Background()
+		conf, err := config.TransformSimpleToClusterConfig(
+			ctx, runtimes.Docker, simpleConfig)
+		Expect(err).NotTo(HaveOccurred())
 
-	kubeconfig, err := client.KubeconfigGet(ctx, runtimes.Docker, &conf.Cluster)
-	Expect(err).NotTo(HaveOccurred())
+		err = client.ClusterRun(ctx, runtimes.Docker, conf)
+		Expect(err).NotTo(HaveOccurred())
 
-	restConfig, err = clientcmd.NewDefaultClientConfig(*kubeconfig, nil).ClientConfig()
-	Expect(err).NotTo(HaveOccurred())
+		kubeconfig, err := client.KubeconfigGet(ctx, runtimes.Docker, &conf.Cluster)
+		Expect(err).NotTo(HaveOccurred())
+
+		restConfig, err = clientcmd.NewDefaultClientConfig(*kubeconfig, nil).ClientConfig()
+		Expect(err).NotTo(HaveOccurred())
+	} else {
+		fmt.Println("KUBECONFIG=" + os.Getenv("KUBECONFIG"))
+	}
 
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
@@ -103,19 +116,28 @@ var _ = BeforeSuite(func() {
 			"../resources",
 		},
 		BinaryAssetsDirectory: "../../testbin/bin",
-		UseExistingCluster:    pointer.Bool(true),
-		Config:                restConfig,
+		UseExistingCluster:    &useExisting,
+		Config:                restConfig, // this will be nil if E2E_USE_EXISTING is set
 		Scheme:                scheme.Scheme,
+		CRDInstallOptions: envtest.CRDInstallOptions{
+			CleanUpAfterUse: useExisting,
+		},
 	}
 
-	_, k8sClient = test.RunTestEnvironment(testEnv,
+	var mgr manager.Manager
+	mgr, k8sClient = test.RunTestEnvironment(testEnv,
 		&demo.OpniDemoReconciler{},
 		&controllers.OpniClusterReconciler{},
 	)
+	if restConfig == nil {
+		restConfig = mgr.GetConfig()
+	}
 })
 
 var _ = AfterSuite(func() {
 	err := testEnv.Stop()
 	Expect(err).NotTo(HaveOccurred())
-	deleteTestClusterIfExists()
+	if !useExisting {
+		deleteTestClusterIfExists()
+	}
 })
